+++ hera-all/milog/log-agent/src/main/resources/config/open.properties
@@ -2,9 +2,9 @@ app_name=milog_agent
 nacosAddr=127.0.0.1:80
 serviceName=milog_manager_server
 
-log.path=/tmp/
+log.path=/home/work/log/log-agent
 # agent\u91C7\u96C6\u8FDB\u5EA6\u5B58\u50A8\u8DEF\u5F84
-agent.memory.path=/tmp/
+agent.memory.path=/home/work/log/log-agent
 #agent.memory.path=D:\work
 # rpc:\u4ECE\u8FDC\u7A0Blog-manager\u8BFB\u53D6 || json: \u4ECE\u672C\u5730agent_channel_config.json\u8BFB\u53D6
-agent.channel.locator=json
+agent.channel.locator=rpc
